[SSCP[llvm-to-amdgpu] Use --offload-arch instead of --cuda-gpu-arch when finding bitcode libraries
It seems that certain `hipcc` versions do not take into account `--cuda-gpu-arch` anymore, which creates an issue if the default target that `hipcc` wants to use does not align with what AdaptiveCpp wants to target.

This PR therefore changes `--cuda-gpu-arch` use to `--offload-arch` which was reported to work fine.

Fixes #1723 

We may have to double check this PR against older ROCm versions to ensure that we don't run into the reverse problem there, i.e. `--offload-arch` not being fully supported yet.